requirements.txt
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
